feat(reviewer): Implement reviewer service (feature/reviewer)
Implement `UserService` and `UserServiceImpl` to handle reviewer
registration, retrieval, and update operations.  Includes
validation to ensure only the authenticated user can modify
their own reviewer profile.  Also updates user role to
`REVIEWER` upon successful registration.
